mdweb.sh: follow the 2-step specification algorithm
[Appendix: A parsing strategy](https://spec.commonmark.org/0.31.2/#appendix-a-parsing-strategy)

Currently, a one-step algorithm is implemented, but it stumbled on
[Example 263](https://spec.commonmark.org/0.31.2/#example-263), where
it is necessary to frame the first element of the list with the
paragraph tag if the list is interrupted by an empty line.

Therefore, at the first stage, a tree of blocks will be created in the
`CONTAINER` array with the depth and types of child blocks in the form
of indexes of this array. The order of indexes is stored in the
`CONTAINER_TREE` array.

At the second stage, you will pass through the tree to print the content
and tags.

Signed-off-by: Semyon A Mironov <s.mironov@atwis.org>